[General]

+++ XClipper.Android/app/src/main/java/com/kpstv/xclipper/extensions/utils/FirebaseUtils.kt
@@ -9,6 +9,7 @@ import com.kpstv.xclipper.data.provider.DBConnectionProvider
 import com.kpstv.xclipper.data.provider.FirebaseProvider
 import com.kpstv.xclipper.data.provider.PreferenceProvider
 import com.kpstv.xclipper.data.repository.MainRepository
+import com.kpstv.hvlog.HVLog
 import com.kpstv.xclipper.extensions.decrypt
 import es.dmoral.toasty.Toasty
 
@@ -25,17 +26,20 @@ class FirebaseUtils(
 
     fun observeDatabaseChangeEvents(): Unit =
         with(context) {
+            if (firebaseProvider.isObservingChanges()) return@with
             if (!App.observeFirebase) return@with
+
+            HVLog.d("Attached")
             firebaseProvider.observeDataChange(
                 changed = {
                     if (App.observeFirebase)
                         repository.updateClip(it?.Clips?.last()?.decrypt())
-//                    Log.e(TAG, "User has changed")
                 },
                 removed = { items -> // Unencrypted listOf data
                     items.forEach { repository.deleteClip(it) }
                 },
                 error = {
+                    HVLog.d()
                     Log.e(TAG, "Error: ${it.message}")
                 },
                 deviceValidated = { isValidated ->
@@ -73,10 +77,12 @@ class FirebaseUtils(
     }
 
     fun observeDatabaseInitialization() {
+        HVLog.d()
         firebaseProvider.isInitialized().observeForever(databaseInitializationObserver)
     }
 
     fun removeDatabaseInitializationObservation() {
+        HVLog.d()
         firebaseProvider.isInitialized().removeObserver(databaseInitializationObserver)
         firebaseProvider.removeDataObservation()
     }
